Version 0.02.00
finished all function of volunteerBL and VolunteersBL

+++ PoliceVolnteerDAL/PoliceVolnteerDAL/CarsReportsDAL.cs
@@ -12,7 +12,7 @@ namespace PoliceVolnteerDAL
 
     public class CarsReportsDAL
     {        
-        public static bool AddCarReport(int cShiftCode, string cCarID, int Distance)
+        public static bool AddCarReport(int cShiftCode, string cCarID, double Distance)
         {
             try
             {
